Task 14: Test suite — Pest conversion and 43 passing tests
Installed Pest v3 with pest-plugin-laravel and converted all feature
tests from PHPUnit class style to Pest DSL. Fixed three service bugs
uncovered during testing: whereDate() for SQLite date column comparisons
in ActivityEventService, AntiCheatService, and StreakFreezeService;
and (int) cast on Carbon::diffInDays() float return in
StreakEvaluationService to fix freeze gap detection.

+++ app/Services/StreakEvaluationService.php
@@ -150,8 +150,9 @@ class StreakEvaluationService
             return false;
         }
 
-        // diffInDays === 2 means exactly 1 missed day between lastCompleted and today
-        $gap = Carbon::parse($lastCompletedDate)->diffInDays(Carbon::parse($localDate));
+        // diffInDays === 2 means exactly 1 missed day between lastCompleted and today.
+        // Cast to int because diffInDays() returns float in recent Carbon versions.
+        $gap = (int) Carbon::parse($lastCompletedDate)->diffInDays(Carbon::parse($localDate));
 
         if ($gap !== 2) {
             return false;
